ingestion/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
google-cloud-pubsub>=2.19.0
pydantic>=2.6.0
python-dotenv>=1.0.0
